refactor: Rename class
